feat: add Editor component in RoleInfoModal.jsx
