+++ src/components/UserProfile/EditableModal/__tests__/roleInfoModal.test.js
@@ -3,7 +3,7 @@ import { render, fireEvent } from '@testing-library/react';
 import { Provider } from 'react-redux'; 
 import configureStore from 'redux-mock-store';
 import { themeMock } from '__tests__/mockStates';
-import RoleInfoModal from '../roleInfoModal';
+import RoleInfoModal from '../RoleInfoModal';
 
 const mockStore = configureStore([]);
